duplication/requirements.txt
requests
python-dotenv
aiokafka
